[Core] Extract BaseLinearSolver class into its own files
